삭제 시, confirm 버튼이 계속 뜨는 현상을 조건문을 바꿨는데 이유는 모르겠다..

+++ src/js/features/handleEditList.js
@@ -10,8 +10,8 @@ const edit = key => {
 };
 
 const del = key => {
-    if (confirm('정말 삭제하시겠습니까?')) {
-        manageMenuList({ type: 'DELETE', key });
+    {
+        confirm('정말 삭제하시겠습니까?') && manageMenuList({ type: 'DELETE', key });
     }
 };
